refactor: Moved static assets from `public` to `assets`
