+++ src/assets/fonts/Lato-Regular.ttf
