+++ src/assets/fonts/index.ts
@@ -1,0 +1,6 @@
+import latoRegularEot from "./Lato-Regular.eot";
+import latoRegularTtf from "./Lato-Regular.ttf";
+import latoRegularWoff from "./Lato-Regular.woff";
+import latoRegularWoff2 from "./Lato-Regular.woff2";
+
+export { latoRegularEot, latoRegularTtf, latoRegularWoff, latoRegularWoff2 };
