+++ src/assets/images/index.ts
@@ -1,0 +1,17 @@
+import AicsLogoAndNameSVGAsset from "./AICS-logo-and-name.svg?react";
+import AicsLogoSVGAsset from "./AICS-logo.svg?react";
+import DropdownSVGAsset from "./dropdown-arrow.svg?react";
+import ExportIconSVGAsset from "./icon-export.svg?react";
+import SwitchIconSVGAsset from "./icon-switch-arrows.svg?react";
+import NoImageSVGAsset from "./no-image.svg?react";
+import SpinBoxHandleDownSVGAsset from "./spin_box-handle-down.svg?react";
+import SpinBoxHandleUpSVGAsset from "./spin_box-handle-up.svg?react";
+
+export const AicsLogoAndNameSVG = AicsLogoAndNameSVGAsset;
+export const AicsLogoSVG = AicsLogoSVGAsset;
+export const DropdownSVG = DropdownSVGAsset;
+export const SwitchIconSVG = SwitchIconSVGAsset;
+export const NoImageSVG = NoImageSVGAsset;
+export const ExportIconSVG = ExportIconSVGAsset;
+export const SpinBoxHandleDownSVG = SpinBoxHandleDownSVGAsset;
+export const SpinBoxHandleUpSVG = SpinBoxHandleUpSVGAsset;
