+++ src/assets/index.ts
@@ -1,13 +1,3 @@
-import AicsLogoSVGAsset from "../assets/AICS-logo.svg?react";
-import AicsLogoAndNameSVGAsset from "../assets/AICS-logo-and-name.svg?react";
-import DropdownSVGAsset from "../assets/dropdown-arrow.svg?react";
-import ExportIconSVGAsset from "../assets/icon-export.svg?react";
-import SwitchIconSVGAsset from "../assets/icon-switch-arrows.svg?react";
-import NoImageSVGAsset from "../assets/no-image.svg?react";
-
-export const AicsLogoAndNameSVG = AicsLogoAndNameSVGAsset;
-export const AicsLogoSVG = AicsLogoSVGAsset;
-export const DropdownSVG = DropdownSVGAsset;
-export const SwitchIconSVG = SwitchIconSVGAsset;
-export const NoImageSVG = NoImageSVGAsset;
-export const ExportIconSVG = ExportIconSVGAsset;
+export * from "./fonts";
+export * from "./images";
+export * from "./videos";
