+++ src/assets/videos/index.ts
@@ -1,0 +1,3 @@
+import bannerVideo from "./banner_video.mp4";
+
+export { bannerVideo };
